UPDATE moved the login post request to home context so that the registration can form can use it as well
