+++ erestaurant-app/src/routes/Home/Home.js
@@ -5,8 +5,11 @@ import LoginForm from "./components/LoginForm";
 import RegForm from "./components/RegForm";
 import Styles from "./Styles";
 import HomeContext from "./HomeContext";
+import queryString from "querystring";
+import Cookies from "universal-cookie";
 
-function Home(props) {
+function Home() {
+  const cookies = new Cookies();
   const [displaySignIn, setDisplaySignIn] = useState(true);
 
   function switchMethod() {
@@ -16,8 +19,39 @@ function Home(props) {
     );
   }
 
+  async function checkLogin(username, password) {
+    const response = await fetch("/Login", {
+      method: "POST",
+      headers: {
+        "Content-type": "application/x-www-form-urlencoded",
+      },
+      body: queryString.stringify({
+        username: username,
+        password: password,
+      }),
+    });
+
+    response
+      .json()
+      .then((res) => {
+        if (res.auth) {
+          const expiryDate = new Date();
+          expiryDate.setMinutes(expiryDate.getMinutes() + 30);
+          console.log(expiryDate);
+
+          cookies.set("Session id", res.session_id, {
+            expires: expiryDate,
+          });
+          return true;
+        } else {
+          return false;
+        }
+      })
+      .catch((err) => console.log(err));
+  }
+
   return (
-    <HomeContext.Provider value={{ displaySignIn, switchMethod }}>
+    <HomeContext.Provider value={{ displaySignIn, switchMethod, checkLogin }}>
       <div className="Home">
         <Helmet>
           <title>Login to your account at DineOut</title>
